pyproject.toml
[tool.poetry]
name = "reddit-finance-webscrapper"
version = "0.1.0"
description = ""
authors = ["monaco89 <nick.monaco15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
black = "^20.8b1"
flake8 = "^3.8.4"
requests = "^2.25.1"
selenium = "^3.141.0"
webdriver-manager = "^3.2.2"
python-dateutil = "^2.8.1"
numpy = "^1.19.5"
pygsheets = "^2.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
